creating document DTOs
