+++ patricare-back/src/modules/document/dto/create-document.dto.ts
@@ -1,0 +1,9 @@
+import { IsString, IsOptional } from 'class-validator';
+
+export class CreateDocumentDto {
+  @IsString() nom: string;
+  @IsString() type: string;
+
+  @IsOptional()
+  dateModification?: Date;
+}
