+++ patricare-back/src/modules/document/dto/update-document.dto.ts
@@ -1,0 +1,7 @@
+import { IsString, IsOptional } from 'class-validator';
+
+export class UpdateDocumentDto {
+  @IsOptional() @IsString() nom?: string;
+  @IsOptional() @IsString() type?: string;
+  @IsOptional() dateModification?: Date;
+}
